api/requirements.txt
absl-py==0.9.0
anvil-uplink==0.3.29
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.7
astor==0.8.1
astroid==2.3.3
asyncio==3.4.3
attrs==19.3.0
autopep8==1.5
backcall==0.1.0
bleach==3.1.1
blis==0.4.1
bqplot==0.11.0
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cycler==0.10.0
cymem==2.0.3
Cython==0.29.16
decorator==4.4.2
deeplift==0.6.10.0
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.4
docopt==0.6.2
docutils==0.16
e==1.4.5
entrypoints==0.3
filelock==3.0.12
fileupload==0.1.5
flake8==3.5.0
Flask==1.1.1
Flask-Cors==3.0.8
future==0.18.2
gast==0.2.2
get==2019.4.13
gevent==1.5.0
google-auth==1.4.2
google-auth-oauthlib==0.4.1
google-colab==1.0.0
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
image==1.5.28
imageio==2.8.0
importlib-metadata==1.5.0
importlib-resources==1.4.0
innvestigate==1.0.8
investigate==1.8.0
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.0
json5==0.9.2
jsonpickle==1.4.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-server==1.0.7
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
more-itertools==8.2.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==4.4.0
nbinteract==0.2.5
networkx==2.4
notebook==5.2.2
numpy==1.18.2
oauthlib==3.1.0
opencv-contrib-python-headless==4.2.0.34
opencv-python==4.1.2.30
opt-einsum==3.2.0
packaging==20.3
pandas==0.25.2
pandocfilters==1.4.2
parameterized==0.7.1
parso==0.6.2
pbr==5.4.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
plac==1.1.3
pluggy==0.13.1
portpicker==1.2.0
post==2019.4.13
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==1.0.18
protobuf==3.11.3
ptyprocess==0.6.0
public==2019.4.13
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.0
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.6.1
pylint==2.4.4
pymongo==3.10.1
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
pytest-timeout==1.3.4
python-dateutil==2.8.1
python-dotenv==0.11.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
query-string==2019.4.13
readme-renderer==24.0
request==2019.4.13
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.1.0
Send2Trash==1.5.0
sentencepiece==0.1.85
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
spacy==2.2.4
sqlparse==0.3.1
srsly==1.0.2
tb-nightly==1.14.0a20190603
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==1.14.0.dev2019060501
Theano==1.0.4
thinc==7.4.0
toolz==0.10.0
torch==1.4.0
torchray==1.0.0.1
torchtext==0.5.0
torchvision==0.5.0
tornado==4.5.3
tqdm==4.43.0
traitlets==4.3.3
traittypes==0.2.1
twine==3.1.1
typed-ast==1.4.1
urllib3==1.24.3
virtualenv==20.0.9
vocab==0.0.4
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
ws4py==0.3.4
zipp==3.1.0
